Bug fix: When reading input Stream in .NET assembly upload fails in 64-bit process, the transfer is not interrupted
Source commit: 5f7060355fcd15b6d610c0701389372f6bd85d55

+++ source/windows/ConsoleRunner.cpp
@@ -976,7 +976,7 @@ void __fastcall TExternalConsole::TransferOut(const unsigned char * Data, size_t
     try
     {
       CommStruct->Event = TConsoleCommStruct::TRANSFEROUT;
-      size_t BlockLen = std::min(Len - Offset, sizeof(CommStruct->TransferEvent.Data));
+      unsigned int BlockLen = std::min(Len - Offset, sizeof(CommStruct->TransferEvent.Data));
       memcpy(CommStruct->TransferEvent.Data, Data + Offset, BlockLen);
       CommStruct->TransferEvent.Len = BlockLen;
       Offset += BlockLen;
